+++ R/SingleGroup-methods.R
@@ -558,7 +558,8 @@ setMethod(
 #' @param type type of residuals to be displayed.
 #'   Can be either \code{'LD'} or \code{'LDG2'} for a local dependence matrix based on the
 #'   X2 or G2 statistics (Chen & Thissen, 1997), \code{'Q3'} for the statistic proposed by
-#'   Yen (1984), or \code{'exp'} for the expected values for the frequencies of every response pattern.
+#'   Yen (1984), \code{'exp'} for the expected values for the frequencies of every response pattern,
+#'   and \code{'expfull'} for the expected values for every theoretically observable response pattern.
 #'   For the 'LD' and 'LDG2' types, the upper diagonal elements represent the standardized
 #'   residuals in the form of signed Cramers V coefficients
 #' @param tables logical; for LD type, return the observed, expected, and standardized residual
@@ -608,6 +609,25 @@ setMethod(
 #' residuals(x, type = 'exp')
 #' residuals(x, suppress = .15)
 #' residuals(x, df.p = TRUE)
+#'
+#' # Pearson's X2 estimate for goodness-of-fit
+#' full_table <- residuals(x, type = 'expfull')
+#' head(full_table)
+#' X2 <- with(full_table, sum((freq - exp)^2 / exp))
+#' df <- nrow(full_table) - extract.mirt(x, 'nest') - 1
+#' p <- pchisq(X2, df = df, lower.tail=FALSE)
+#' data.frame(X2, df, p, row.names='Pearson-X2')
+#'
+#' # above FOG test as a function
+#' PearsonX2 <- function(x){
+#'    full_table <- residuals(x, type = 'expfull')
+#'    X2 <- with(full_table, sum((freq - exp)^2 / exp))
+#'    df <- nrow(full_table) - extract.mirt(x, 'nest') - 1
+#'    p <- pchisq(X2, df = df, lower.tail=FALSE)
+#'    data.frame(X2, df, p, row.names='Pearson-X2')
+#' }
+#' PearsonX2(x)
+#'
 #'
 #' # extract results manually
 #' out <- residuals(x, df.p = TRUE, verbose=FALSE)
@@ -770,6 +790,42 @@ setMethod(
                 class(tabdata) <- c('mirt_df', 'data.frame')
                 return(tabdata)
             }
+        } else if(type == 'expfull'){
+            K <- extract.mirt(object, 'K')
+            nitems <- length(K)
+            resp <- vector('list', nitems)
+            for(i in seq_len(nitems))
+                resp[[i]] <- 0L:(K[i]-1L)
+            tabdata <- expand.grid(resp)
+            rownames(tabdata) <- NULL
+            tabdata <- t(t(tabdata) + extract.mirt(object, 'mins'))
+            colnames(tabdata) <- colnames(extract.mirt(object, 'data'))
+            sv <- mod2values(object)
+            itemtype <- extract.mirt(object, 'itemtype')
+            nfact <- extract.mirt(object, 'nfact')
+            full_object <- mirt(tabdata, nfact, itemtype=itemtype, pars=sv, TOL=NaN)
+            Pl <- full_object@Internals$Pl
+            r <- integer(length(Pl))
+            ro <- object@Data$Freq[[1L]]
+            N <- sum(ro)
+            for(i in seq_len(length(ro))){
+                pick <- colSums(t(tabdata) == object@Data$tabdata[i,]) == ncol(tabdata)
+                r[pick] <- ro[i]
+            }
+            res <- (r - Pl * N) / sqrt(Pl * N)
+            expected <- N * Pl
+            ISNA <- is.na(rowSums(tabdata))
+            expected[ISNA] <- res[ISNA] <- NA
+            tabdata <- data.frame(tabdata,r,expected,res)
+            colnames(tabdata) <- c(colnames(object@Data$data),"freq","exp","res")
+            tabdata <- tabdata[do.call(order, as.data.frame(tabdata[,1:J])),]
+            rownames(tabdata) <- 1:length(r)
+            if(!is.null(printvalue)){
+                if(!is.numeric(printvalue)) stop('printvalue is not a number.', call.=FALSE)
+                tabdata <- tabdata[abs(tabdata[ ,ncol(tabdata)]) > printvalue, ]
+            }
+            class(tabdata) <- c('mirt_df', 'data.frame')
+            return(tabdata)
         } else if(type == 'Q3'){
             if(discrete && !use_dentype_estimate)
                 stop('residual type not supported for discrete density forms', call.=FALSE)